test: refactoring setup data settings
