+++ src/test/java/project/linkarchive/backend/util/setUpData/LinkSetUpData.java
@@ -5,19 +5,9 @@ import project.linkarchive.backend.link.domain.Link;
 import project.linkarchive.backend.link.request.CreateLinkRequest;
 import project.linkarchive.backend.user.domain.User;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import static project.linkarchive.backend.util.constant.Constants.*;
 
-public class LinkSetUpData {
-
-    public User user;
-    public Link link;
-
-    public List<String> tags = new ArrayList<>();
-
-    public CreateLinkRequest createLinkRequest;
+public class LinkSetUpData extends SetUpData {
 
     @BeforeEach
     public void setup() {
